Renamed `view.dart` to `chart_view.dart`
