+++ lib/src/guide/interaction/crosshair.dart
@@ -1,19 +1,19 @@
 import 'dart:ui';
 
-import 'package:graphic/src/graffiti/element/arc.dart';
-import 'package:graphic/src/graffiti/element/line.dart';
-import 'package:graphic/src/graffiti/element/rect.dart';
-import 'package:graphic/src/util/collection.dart';
 import 'package:graphic/src/chart/chart.dart';
-import 'package:graphic/src/chart/view.dart';
+import 'package:graphic/src/chart/chart_view.dart';
 import 'package:graphic/src/common/operators/render.dart';
 import 'package:graphic/src/coord/coord.dart';
 import 'package:graphic/src/coord/polar.dart';
 import 'package:graphic/src/coord/rect.dart';
 import 'package:graphic/src/dataflow/tuple.dart';
+import 'package:graphic/src/graffiti/element/arc.dart';
 import 'package:graphic/src/graffiti/element/element.dart';
+import 'package:graphic/src/graffiti/element/line.dart';
+import 'package:graphic/src/graffiti/element/rect.dart';
 import 'package:graphic/src/graffiti/scene.dart';
 import 'package:graphic/src/interaction/selection/selection.dart';
+import 'package:graphic/src/util/collection.dart';
 
 /// The specification of a crosshair
 ///
