+++ lib/src/parse/parse.dart
@@ -2,19 +2,11 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:flutter/painting.dart';
-import 'package:graphic/src/common/builtin_layers.dart';
-import 'package:graphic/src/encode/encode.dart';
-import 'package:graphic/src/encode/channel.dart';
-import 'package:graphic/src/encode/color.dart';
-import 'package:graphic/src/encode/gradient.dart';
-import 'package:graphic/src/encode/position.dart';
-import 'package:graphic/src/encode/shape.dart';
-import 'package:graphic/src/encode/elevation.dart';
-import 'package:graphic/src/encode/size.dart';
 import 'package:graphic/src/algebra/varset.dart';
 import 'package:graphic/src/chart/chart.dart';
+import 'package:graphic/src/chart/chart_view.dart';
 import 'package:graphic/src/chart/size.dart';
-import 'package:graphic/src/chart/view.dart';
+import 'package:graphic/src/common/builtin_layers.dart';
 import 'package:graphic/src/common/defaults.dart';
 import 'package:graphic/src/common/dim.dart';
 import 'package:graphic/src/common/label.dart';
@@ -25,9 +17,16 @@ import 'package:graphic/src/coord/polar.dart';
 import 'package:graphic/src/coord/rect.dart';
 import 'package:graphic/src/data/data_set.dart';
 import 'package:graphic/src/dataflow/operator.dart';
+import 'package:graphic/src/dataflow/tuple.dart';
+import 'package:graphic/src/encode/channel.dart';
+import 'package:graphic/src/encode/color.dart';
+import 'package:graphic/src/encode/elevation.dart';
+import 'package:graphic/src/encode/encode.dart';
+import 'package:graphic/src/encode/gradient.dart';
+import 'package:graphic/src/encode/position.dart';
+import 'package:graphic/src/encode/shape.dart';
+import 'package:graphic/src/encode/size.dart';
 import 'package:graphic/src/graffiti/element/element.dart';
-import 'package:graphic/src/mark/mark.dart';
-import 'package:graphic/src/mark/modifier/modifier.dart';
 import 'package:graphic/src/guide/annotation/custom.dart';
 import 'package:graphic/src/guide/annotation/element.dart';
 import 'package:graphic/src/guide/annotation/line.dart';
@@ -36,15 +35,16 @@ import 'package:graphic/src/guide/annotation/tag.dart';
 import 'package:graphic/src/guide/axis/axis.dart';
 import 'package:graphic/src/guide/interaction/crosshair.dart';
 import 'package:graphic/src/guide/interaction/tooltip.dart';
+import 'package:graphic/src/interaction/event.dart';
 import 'package:graphic/src/interaction/gesture.dart';
 import 'package:graphic/src/interaction/selection/interval.dart';
 import 'package:graphic/src/interaction/selection/point.dart';
 import 'package:graphic/src/interaction/selection/selection.dart';
-import 'package:graphic/src/interaction/event.dart';
+import 'package:graphic/src/mark/mark.dart';
+import 'package:graphic/src/mark/modifier/modifier.dart';
 import 'package:graphic/src/scale/linear.dart';
 import 'package:graphic/src/scale/ordinal.dart';
 import 'package:graphic/src/scale/scale.dart';
-import 'package:graphic/src/dataflow/tuple.dart';
 import 'package:graphic/src/scale/time.dart';
 import 'package:graphic/src/shape/shape.dart';
 import 'package:graphic/src/variable/transform/filter.dart';
